Moved PctString to separate file
