+++ modules/core/src/main/scala/ch/epfl/bluebrain/nexus/rdf/Iri.scala
@@ -6,12 +6,12 @@ import cats.{Eq, Show}
 import ch.epfl.bluebrain.nexus.rdf.Iri.Host.{IPv4Host, IPv6Host, NamedHost}
 import ch.epfl.bluebrain.nexus.rdf.Iri.Path._
 import ch.epfl.bluebrain.nexus.rdf.Iri._
+import ch.epfl.bluebrain.nexus.rdf.PctString._
 import org.parboiled2.ErrorFormatter
 import org.parboiled2.Parser.DeliveryScheme.{Either => E}
-import java.nio.charset.StandardCharsets.UTF_8
+
 import scala.collection.immutable.SortedSet
 import scala.collection.{immutable, SeqView, SortedMap}
-import org.parboiled2.CharUtils._
 
 /**
   * An Iri as defined by RFC 3987.
@@ -1019,36 +1019,4 @@ object Iri {
       case url: Url       => url.show
       case urn: Urn       => urn.show
     }
-
-  private val `sub-delims` = Set('!', '$', '&', '\'', '(', ')', '*', '+', ',', ';', '=').map(_.toByte)
-  private val unreserverd = Set('-', '_', '.', '~').map(_.toByte) ++ ('a' to 'z').map(_.toByte).toSet ++ ('A' to 'Z')
-    .map(_.toByte)
-    .toSet ++ ('0' to '9').map(_.toByte).toSet
-  private val pchar = `sub-delims` ++ unreserverd + ':'.toByte + '@'.toByte
-
-  private def pctEncode(s: String) = {
-    val in = s.getBytes(UTF_8)
-    in.foldLeft(new StringBuilder()) {
-        case (sb, b) if pchar.contains(b) => sb.append(b.toChar)
-        case (sb, b)                      => sb.append("%").append(upperHexDigit((b >> 4) & 0xF)).append(upperHexDigit(b & 0xF))
-      }
-      .toString()
-  }
-
-  private[rdf] trait PctString {
-    def value: String
-
-    /**
-      * @return the UTF-8 representation
-      */
-    lazy val asString: String =
-      value
-
-    /**
-      * @return the string representation using percent-encoding for any character that
-      *         is not contained in the Set ''pchar''
-      */
-    lazy val pctEncoded: String =
-      pctEncode(value)
-  }
 }
