+++ modules/core/src/test/scala/ch/epfl/bluebrain/nexus/rdf/RelativeIriSpec.scala
@@ -1,7 +1,6 @@
 package ch.epfl.bluebrain.nexus.rdf
 
 import cats.kernel.Eq
-//import cats.syntax.show._
 import ch.epfl.bluebrain.nexus.rdf.Iri._
 import org.scalatest.{EitherValues, Inspectors, Matchers, WordSpecLike}
